requirements.txt
wandb
pandas
numpy
tensorflow
soundfile
librosa
tqdm
matplotlib
Pillow
scikit-learn